[Edited]: แก้ return type จาก int เป็น uint64 ในฟังก์ชัน convertToSecond

+++ src/caldate/caldate.go
@@ -1,5 +1,5 @@
 package caldate
 
-func convertToSecond(days int) int {
-	return days * 86400
+func convertToSecond(days int) uint64 {
+	return uint64(days * 86400)
 }
